requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
statsmodels>=0.14.0
numba>=0.57.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pyyaml>=6.0
tqdm>=4.65.0
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.25.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
